Add better file validation

+++ src/Steps/CopyingAsarInstall.ts
@@ -10,12 +10,12 @@ import * as helpers from '../utils/FileHelper';
 export async function execute(options: Options): Promise<boolean> {
     const tidalPath = options.overwritePath || await getTidalDirectory();
     
-    if (!tidalPath || !fs.existsSync(tidalPath)) {
+    if (!tidalPath || !await fs.existsSync(tidalPath)) {
         msg.stepError(Steps.COPYING_ASAR_INSTALL, 'Tidal is not installed', new Error('Invalid File Path'));
         return false;
     }
     
-    const asarFiles = fs.readdirSync(tidalPath).filter(file => file.endsWith('.asar'));
+    const asarFiles = await fs.readdirSync(tidalPath).filter(file => file.endsWith('.asar'));
     if (asarFiles.length === 0) {
         msg.stepError(Steps.COPYING_ASAR_INSTALL, 'Tidal is not installed or your installation is corrupt!', new Error('Asar file missing'));
         return false;
@@ -23,15 +23,20 @@ export async function execute(options: Options): Promise<boolean> {
 
     msg.stepLog(Steps.COPYING_ASAR_INSTALL, 'Copying app.asar to original.asar');
     let originalAsarPath = path.join(tidalPath, 'original.asar');
-    const originalAsarExists = fs.existsSync(originalAsarPath);
+    const appAsarPath = path.join(tidalPath, 'app.asar');
+    const originalAsarExists = await fs.existsSync(originalAsarPath);
     if (!originalAsarExists) {
+        const appAsarExists = await fs.existsSync(appAsarPath);
+        if (!appAsarExists) {
+            msg.stepError(Steps.COPYING_ASAR_INSTALL, 'app.asar not found. Your installation is corrupt! Please reinstall tidal!', new Error('app.asar not found'));
+            return false;
+        }
         msg.stepLog(Steps.COPYING_ASAR_INSTALL, 'Creating original.asar backup');
         await fs.copyFileSync(originalAsarPath, path.join(tidalPath, 'app.asar'));
     }
 
-    const appAsarPath = path.join(tidalPath, 'app.asar');
-    if (fs.existsSync(appAsarPath)) {
-        fs.unlinkSync(appAsarPath);
+    if (await fs.existsSync(appAsarPath)) {
+        await fs.unlinkSync(appAsarPath);
     }
 
     msg.stepLog(Steps.COPYING_ASAR_INSTALL, 'Copying app.asar to original.asar completed successfully');
